Fix metap output to include <?> on missing types.
