+++ tests/jvm/src/test/resources/metacp.expect
@@ -1792,8 +1792,8 @@ types.Test.Literal.float(). => final val method float: : 1.0f
 types.Test.Literal.float. => private[this] final val field float: 1.0f
 types.Test.Literal.int(). => final val method int: : 1
 types.Test.Literal.int. => private[this] final val field int: 1
-types.Test.Literal.javaEnum(?). => final val method javaEnum
-types.Test.Literal.javaEnum. => private[this] final val field javaEnum
+types.Test.Literal.javaEnum(?). => final val method javaEnum<?>
+types.Test.Literal.javaEnum. => private[this] final val field javaEnum<?>
 types.Test.Literal.long(). => final val method long: : 1L
 types.Test.Literal.long. => private[this] final val field long: 1L
 types.Test.Literal.nil(). => final val method nil: : null
